Move parsing into separate file
